lado server exercises más o menos acabado
