+++ src/java/entities/CallType.java
@@ -3,14 +3,16 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
 /**
  *
- * @author 2dam
+ * @author Nerea
  */
-public class Enrolled {
-    private Subject subject;
-    private Student student;
-    private Boolean isMatriculate;
+public enum CallType {
+    FIRST,
+    SECOND,
+    THIRD,
+    FOURTH,
+    FIFTH,
 }
