+++ src/java/entities/Enrolled.java
@@ -3,14 +3,14 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
 /**
  *
- * @author Nerea
+ * @author 2dam
  */
-public enum LanguageType {
-    ENGLISH,
-    SPANISH,
-    BASQUE
+public class Enrolled {
+    private Subject subject;
+    private Student student;
+    private Boolean isMatriculate;
 }
