+++ src/java/entities/Exam.java
@@ -3,10 +3,11 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import java.io.File;
 import java.io.Serializable;
 import java.util.Date;
 import java.util.Set;
@@ -18,98 +19,100 @@ import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.xml.bind.annotation.XmlRootElement;
 
-
 /**
  *
- * @author Nerea
+ * @author 2dam
  */
 @Entity
-@Table(name = "unit", schema = "bytebuddiesbd")
+@Table(name = "exam", schema = "bytebuddiesbd")
 @XmlRootElement
-public class Unit implements Serializable{
+public class Exam implements Serializable {
+
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Integer id;
-    private String name;
     private String description;
     @Temporal(javax.persistence.TemporalType.DATE)
     @JsonSerialize(as = Date.class)
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXXX")
     private Date dateInit;
-    @Temporal(javax.persistence.TemporalType.DATE)
-    @JsonSerialize(as = Date.class)
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXXX")
-    private Date dateEnd;
-    private Integer hours;
-    private Set<Exercise> exercises;
+    private Integer duration;
+    private File file;
     private Subject subject;
-    
+    private Set<Mark> marks;
+
     //Setters and Getters
     public Integer getId() {
         return id;
     }
+
     public void setId(Integer id) {
         this.id = id;
     }
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
+
     public String getDescription() {
         return description;
     }
+
     public void setDescription(String description) {
         this.description = description;
     }
+
     public Date getDateInit() {
         return dateInit;
     }
+
     public void setDateInit(Date dateInit) {
         this.dateInit = dateInit;
     }
-    public Date getDateEnd() {
-        return dateEnd;
+
+    public Integer getDuration() {
+        return duration;
     }
-    public void setDateEnd(Date dateEnd) {
-        this.dateEnd = dateEnd;
+
+    public void setDuration(Integer duration) {
+        this.duration = duration;
     }
-    public Integer getHours() {
-        return hours;
+
+    public File getFile() {
+        return file;
     }
-    public void setHours(Integer hours) {
-        this.hours = hours;
+
+    public void setFile(File file) {
+        this.file = file;
     }
-    public Set<Exercise> getExercises() {
-        return exercises;
-    }
-    public void setExercises(Set<Exercise> exercises) {
-        this.exercises = exercises;
-    }
+
     public Subject getSubject() {
         return subject;
     }
+
     public void setSubject(Subject subject) {
         this.subject = subject;
     }
-    
-    //Constructors
-    public Unit(Integer id, String name, String description, Date dateInit, Date dateEnd, Integer hours, Set<Exercise> exercises, Subject subject) {
+
+    public Set<Mark> getMarks() {
+        return marks;
+    }
+
+    public void setMarks(Set<Mark> marks) {
+        this.marks = marks;
+    }
+
+    //Constructor
+    public Exam(Integer id, String description, Date dateInit, Integer duration, File file, Subject subject, Set<Mark> marks) {
         this.id = id;
-        this.name = name;
         this.description = description;
         this.dateInit = dateInit;
-        this.dateEnd = dateEnd;
-        this.hours = hours;
-        this.exercises = exercises;
+        this.duration = duration;
+        this.file = file;
         this.subject = subject;
+        this.marks = marks;
     }
-    public Unit() {
-    }   
+    public Exam() {
+    }
     
-      @Override
+    @Override
     public int hashCode() {
         int hash = 0;
         hash += (id != null ? id.hashCode() : 0);
@@ -119,10 +122,11 @@ public class Unit implements Serializable{
     @Override
     public boolean equals(Object object) {
         // TODO: Warning - this method won't work in the case the id fields are not set
-        if (!(object instanceof Unit)) {
+        if (!(object instanceof Exam)) {
             return false;
         }
-        Unit other = (Unit) object;
+        Exam other = (Exam) object;
         return !((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id)));
     }
+
 }
