+++ src/java/entities/User.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
 import java.io.Serializable;
 import java.util.Date;
